Wyodrębnienie klas do plików
