+++ Main.cpp
@@ -1,133 +1,9 @@
 #include <SFML/Graphics.hpp>
+#include "Player.h"
+#include "Button.h"
 #include <iostream>
 
 enum GameState { MainMenu, OptionsMenu, Gameplay, Pause };
-class Player {
-public:
-    Player(const sf::Vector2f& size, const sf::Vector2f& position, const sf::Color& color) {
-        player.setSize(size);
-        player.setPosition(position);
-        player.setFillColor(color);
-        velocity = sf::Vector2f(0.f, 0.f);
-        isJumping = false;
-        isHoldingJump = false;
-        jumpHoldTime = 0.f;
-    }
-
-    // Copy assignment operator
-    Player& operator=(const Player& other) {
-        if (this != &other) {
-            player = other.player;
-            velocity = other.velocity;
-            isJumping = other.isJumping;
-            isHoldingJump = other.isHoldingJump;
-            jumpHoldTime = other.jumpHoldTime;
-        }
-        return *this;
-    }
-
-    void handleInput(float deltaTime) {
-        // Ruch w lewo i prawo
-        if (sf::Keyboard::isKeyPressed(sf::Keyboard::Left)) {
-            velocity.x = -moveSpeed;
-        } else if (sf::Keyboard::isKeyPressed(sf::Keyboard::Right)) {
-            velocity.x = moveSpeed;
-        } else {
-            velocity.x = 0.f;
-        }
-
-        // Skakanie - wykrywanie przytrzymania klawisza
-        if (sf::Keyboard::isKeyPressed(sf::Keyboard::Space) || sf::Keyboard::isKeyPressed(sf::Keyboard::Up)) {
-            if (!isJumping) {
-                velocity.y = -baseJumpSpeed;
-                isJumping = true;
-                isHoldingJump = true;
-                jumpHoldTime = 0.f;
-            } else if (isHoldingJump && jumpHoldTime < maxHoldTime) {
-                jumpHoldTime += deltaTime;
-                velocity.y = std::max(velocity.y - jumpBoost * deltaTime, -maxJumpSpeed);
-            }
-        } else {
-            isHoldingJump = false;
-        }
-    }
-
-    void update(float deltaTime) {
-        // Grawitacja
-        velocity.y += gravity * deltaTime;
-
-        // Poruszanie graczem
-        player.move(velocity * deltaTime);
-
-        // Ograniczenie pozycji do "podłogi"
-        if (player.getPosition().y + player.getSize().y >= groundHeight) {
-            player.setPosition(player.getPosition().x, groundHeight - player.getSize().y);
-            velocity.y = 0.f;
-            isJumping = false;
-        }
-    }
-
-    void draw(sf::RenderWindow& window) {
-        window.draw(player);
-    }
-
-private:
-    sf::RectangleShape player;
-    sf::Vector2f velocity;
-    bool isJumping;
-    bool isHoldingJump;
-    float jumpHoldTime;
-
-    const float moveSpeed = 200.f;
-    const float baseJumpSpeed = 250.f; // Bazowa prędkość skoku
-    const float maxJumpSpeed = 500.f; // Maksymalna prędkość w górę
-    const float gravity = 981.f;
-    const float groundHeight = 600.f;
-    const float maxHoldTime = 1.0f; // Maksymalny czas trzymania klawisza (sekundy)
-    const float jumpBoost = 500.f; // Dodatkowe przyspieszenie przy trzymaniu klawisza
-};
-
-class Button {
-public:
-    Button(const sf::Vector2f& size, const sf::Vector2f& position, const sf::String& text, const sf::Font& font) {
-        button.setSize(size);
-        button.setPosition(position);
-        button.setFillColor(normalColor);
-
-        buttonText.setFont(font); 
-        buttonText.setString(text);
-        buttonText.setCharacterSize(50);
-        buttonText.setFillColor(sf::Color::White);
-        buttonText.setPosition(
-            position.x + (size.x - buttonText.getLocalBounds().width) / 2,
-            position.y + (size.y - buttonText.getLocalBounds().height) / 2
-        );
-    }
-
-    void draw(sf::RenderWindow& window) {
-        window.draw(button);
-        window.draw(buttonText);
-    }
-
-    void update(const sf::Vector2i& mousePos) {
-        if (button.getGlobalBounds().contains(static_cast<sf::Vector2f>(mousePos))) {
-            button.setFillColor(hoverColor);
-        } else {
-            button.setFillColor(normalColor);
-        }
-    }
-
-    bool isClicked(const sf::Vector2i& mousePos, sf::Event::MouseButtonEvent mouseEvent) {
-        return button.getGlobalBounds().contains(static_cast<sf::Vector2f>(mousePos)) &&
-               mouseEvent.button == sf::Mouse::Left;
-    }
-
-private:
-    sf::RectangleShape button;
-    sf::Text buttonText;
-    sf::Color normalColor = sf::Color::Blue;
-    sf::Color hoverColor = sf::Color::Green;
-};
 
 int main() {
     // Tworzenie okna
